refactor(api): extract task-queue handlers from system.rs (#3749 9/N)
Move /tasks, /tasks/status, /tasks/list, /tasks/{id}, /tasks/{id}/retry
handlers into a new routes::task_queue module. Public paths and JSON
shapes are unchanged; the new router is merged into v1 alongside the
existing system router.

Refs #3749
