+++ crates/librefang-api/src/routes/mod.rs
@@ -33,6 +33,7 @@ pub mod prompts;
 pub mod providers;
 pub mod skills;
 pub mod system;
+pub mod task_queue;
 pub mod terminal;
 pub mod tools_sessions;
 pub mod users;
@@ -68,6 +69,7 @@ pub use plugins::*;
 pub use providers::*;
 pub use skills::*;
 pub use system::*;
+pub use task_queue::*;
 pub use terminal::*;
 pub use tools_sessions::*;
 pub use users::*;
